split errors.Join support
