+++ errors_join_go120_test.go
@@ -1,0 +1,23 @@
+//go:build go1.20
+// +build go1.20
+
+package errors_test
+
+import (
+	"errors"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestJoinFunction(t *testing.T) {
+	t.Parallel()
+
+	err1 := errors.New("foo")
+	err2 := errors.New("bar")
+	err3 := error(nil)
+
+	err := errors.Join(err1, err2, err3)
+
+	require.EqualError(t, err, "foo\nbar")
+}
