+++ errors_test.go
@@ -149,15 +149,3 @@ func TestAsFunction_wrappingError(t *testing.T) {
 		assert.Nil(t, cerr)
 	})
 }
-
-func TestJoinFunction(t *testing.T) {
-	t.Parallel()
-
-	err1 := errors.New("foo")
-	err2 := errors.New("bar")
-	err3 := error(nil)
-
-	err := errors.Join(err1, err2, err3)
-
-	require.EqualError(t, err, "foo\nbar")
-}
